make packages structure uniform
