+++ packages/dynamodb-event-storage-adapter/src/adapter.ts
