+++ packages/dynamodb-event-storage-adapter/src/index.ts
@@ -1,3 +1,3 @@
-export * from './dynamoDb';
+export * from './adapter';
 export { formatEventForTransaction } from './utils/formatEventForTransaction';
 export { pushEventsTransaction } from './utils/pushEventsTransaction';
