+++ packages/dynamodb-event-storage-adapter/src/utils/pushEventsTransaction.unit.test.ts
@@ -7,7 +7,7 @@ import { mockClient } from 'aws-sdk-client-mock';
 import type { EventStoreEventsDetails } from '@castore/core';
 import { counterEventStore, userEventStore } from '@castore/demo-blueprint';
 
-import { DynamoDbEventStorageAdapter } from '../dynamoDb';
+import { DynamoDbEventStorageAdapter } from '../adapter';
 import { formatEventForTransaction } from './formatEventForTransaction';
 import { pushEventsTransaction } from './pushEventsTransaction';
