+++ packages/event-bridge-message-bus-adapter/src/adapter.ts
