+++ packages/in-memory-message-bus-adapter/src/index.ts
@@ -1,2 +1,2 @@
-export { InMemoryMessageBusAdapter } from './inMemory';
+export { InMemoryMessageBusAdapter } from './adapter';
 export type { InMemoryMessageBusMessage } from './message';
