+++ packages/in-memory-message-bus-adapter/src/types.ts
@@ -1,4 +1,4 @@
-import type { EventEmitter } from 'node:events';
+import type { EventEmitter } from 'events';
 
 import type {
   Message,
