+++ packages/inmemory-event-storage-adapter/src/adapter.ts
