+++ packages/inmemory-event-storage-adapter/src/index.ts
@@ -1,1 +1,1 @@
-export * from './inMemory';
+export * from './adapter';
