+++ packages/json-schema-command/src/index.ts
@@ -1,2 +1,2 @@
-export { JSONSchemaCommand } from './jsonSchema';
-export type { OnEventAlreadyExistsCallback } from './jsonSchema';
+export { JSONSchemaCommand } from './command';
+export type { OnEventAlreadyExistsCallback } from './command';
