+++ packages/json-schema-event/src/index.ts
@@ -1,1 +1,1 @@
-export { JSONSchemaEventType } from './jsonSchema';
+export { JSONSchemaEventType } from './eventType';
