+++ packages/redux-event-storage-adapter/src/adapter.ts
